Adjust the accuracy of the Benchmark test, giving the result as a floating point

+++ Tests/Core/Benchmark/Sources/app.h
@@ -44,7 +44,7 @@ private:
 	void write_result();
 	void test();
 	void draw_info(const std::string &text);
-	int run_test();
+	float run_test();
 
 private:
 	bool quit;
